+++ src/components/BaseComponent.php
@@ -1,0 +1,12 @@
+<?php
+namespace VovanVE\HtmlTemplate\components;
+
+use VovanVE\HtmlTemplate\base\BaseObject;
+
+/**
+ * Component to render custom markup
+ * @since 0.1.0
+ */
+abstract class BaseComponent extends BaseObject implements ComponentInterface
+{
+}