+++ tests/unit/compile/cases/html/10.component-single-props.t
@@ -1,0 +1,5 @@
+Lorem <TestComponent foo=42 bar={lol}/> ipsum
+---- CODE ----
+('Lorem ' . ($runtime->createComponent('TestComponent',['foo'=>'42','bar'=>($runtime->param('lol'))])) . ' ipsum')
+---- RESULT ----
+Lorem <!-- Test Component: foo="42" bar="[value of &lol]" /--> ipsum