+++ tests/unit/compile/cases/html/11.component-block-empty-no-props.t
@@ -1,0 +1,5 @@
+Lorem <TestComponent></TestComponent> ipsum
+---- CODE ----
+('Lorem ' . ($runtime->createComponent('TestComponent',[],[])) . ' ipsum')
+---- RESULT ----
+Lorem <!-- Test Component: foo=null bar=null --><!-- /Test Component --> ipsum